Hide and show 'Add to Cart' button
- Added functionality to hide the 'Add to Cart' button once the item has
been added to a user's cart.
- Added functionality to re-show the 'Add to Cart' button once the item
has been deleted from the user's cart, allowing them to add again if
needed.

-  Still having problem. The 'Add to Cart' button adds the specific
item to the user's cart and the 'Add to Cart' button for that specific
product then disappears, not allowing the user to add the same item to
the cart twice.  However, the item stays in the user's cart upon
sign-out. Therefore, when the user signs in, the button is automatically
visible, which allows the user to add multiple of the same item into the
cart. Still working on a fix.